Updated native code
